Statistics: make primitive statistics thread-safe.
git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@24458 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c

+++ src/org/sosy_lab/cpachecker/cpa/smg/refiner/SMGInterpolationTree.java
@@ -455,7 +455,7 @@ public class SMGInterpolationTree {
    *
    * @param file file the file to write to
    */
-  public void exportToDot(PathTemplate file, int refinementCounter) {
+  public void exportToDot(PathTemplate file, long refinementCounter) {
     StringBuilder result = new StringBuilder().append("digraph tree {" + "\n");
     for (Map.Entry<ARGState, ARGState> current : successorRelation.entries()) {
       if (interpolants.containsKey(current.getKey())) {
